Move files around so that we don't break the node build.
