+++ node/src/main/scripts/app.spec.ts
@@ -27,7 +27,7 @@ expect.extend({ toMatchImageSnapshot });
 
 test("videoCut headers", async (done) => {
   const response = await request(underTest).get(
-    "/static/social-header/~cut:(2007.9000244140625,2046.1099853515625),fact:E74aoUY=31c55b20cc10bdfe9b10ce1fbc3d9b3f3bf01098,kind:videoCut"
+    "/static/social-header/~cut:(548.906982421875,551.8199951171875),fact:E74aoUY=8ab2602f03abece68fb5d9cc97e48e2b84b568c1,kind:videoCut"
   );
   expect(response.statusCode).toBe(200);
   expect(response.type).toBe("text/plain");
